+++ src/App.tsx
@@ -4,7 +4,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 
-import ColorThemeProvider from '@components/common/ColorThemeProvider';
+import ColorThemeProvider from '@components/wrapper/ColorThemeProvider';
 import StyledNavigationContainer from '@navigation/_components/styledNavigationContainer/StyledNavigationContainer';
 import { RootNavigation } from '@navigation/RootNavigation';
