+++ src/components/icon/WeatherIcon.tsx
@@ -7,7 +7,7 @@ import ScatteredCloudsSvg from '@assets/svg/weather/WeatherScatteredClouds.svg';
 import ShowerRainSvg from '@assets/svg/weather/WeatherShowerRain.svg';
 import SnowSvg from '@assets/svg/weather/WeatherSnow.svg';
 import ThunderstormSvg from '@assets/svg/weather/WeatherThunderstorm.svg';
-import Show from '@components/common/Show';
+import Show from '@components/wrapper/Show';
 
 import type { IconProps } from '@components/icon/Icon.type';
 import type { WeatherIcon } from '@services/openWeatherOneCall/axios.type';
